WIP: Remove redundant cell styling hook
